Organize checkpoint API sources
